# Some optimizations to Collection::keyBy, Arr::undot, Arr::set, Arr::last and indirectly Arr::keyBy due to Collection::keyBy