--- a/src/Illuminate/Collections/Collection.php
+++ b/src/Illuminate/Collections/Collection.php
@@ -802,19 +802,6 @@ class Collection implements ArrayAccess, CanBeEscapedWhenCastToString, Enumerabl
      */
     public function pluck($value, $key = null)
     {
-        // Fast path for simple value keys (no dots)
-        if (is_string($value) && ! str_contains($value, '.') && is_null($key)) {
-            return new static(array_map(function ($item) use ($value) {
-                if (Arr::accessible($item) && Arr::exists($item, $value)) {
-                    return $item[$value];
-                } elseif (is_object($item) && isset($item->{$value})) {
-                    return $item->{$value};
-                }
-
-                return data_get($item, $value);
-            }, $this->items));
-        }
-
         return new static(Arr::pluck($this->items, $value, $key));
     }