## Performance Optimizations for some Arr and Collection Methods

This PR introduces performance optimizations for several utility methods within `Illuminate\Support\Arr` and `Illuminate\Support\Collection` to improve their efficiency, particularly in common use cases or when used within loops.

### Changes Summary:

1.  **`Arr::last`**: Replaced the previous implementation, which used `array_reverse()` followed by `Arr::first()` when a callback was provided, with a direct backward iteration using `end()` and `prev()`. This avoids the overhead of reversing the entire array.

2.  **`Arr::set`**: Optimized the internal loop used for traversing dot-notation keys by replacing `explode()` with `strtok()`. This offers a minor performance improvement by avoiding the creation of an intermediate array of key segments.

3.  **`Arr::undot`**: Significantly refactored to inline the logic previously handled by repeated calls to `Arr::set`. The key parsing within the inlined logic now uses `strtok()` instead of `explode()`. This dramatically reduces function call overhead and improves key parsing efficiency, especially for arrays with many keys.

4.  **`Collection::keyBy`**: Implemented a specialized fast path for the common case where `$keyBy` is a simple string without dot notation. This path uses direct array/object property access where possible, falling back to `data_get()` only when necessary for complex item types or missing keys. For other `$keyBy` types (callables, dot-notation strings, arrays), it uses the existing `valueRetriever()` method, ensuring compatibility.

### Benchmark Results:

The following benchmark was run using PHP 8.3.20 with **100,000** iterations for each test case.

*(Note: Times are approximate and may vary slightly between runs)*

| Method / Case                      | Optimized Time (s) | Original Time (s) | Improvement          | Notes                                            |
| :--------------------------------- | :----------------- | :---------------- | :------------------- | :----------------------------------------------- |
| **`Arr::last` (callback)**         | ~0.0405            | ~0.1151           | **~2.8x Faster**     | Avoids `array_reverse`.                          |
| **`Arr::set` (dot key)**           | ~0.0445            | ~0.0481           | **~1.1x Faster**     | Minor gain from `strtok` vs `explode`.           |
| **`Arr::undot`**                   | ~0.2313            | ~0.2821           | **~1.2x Faster**     | Inlined logic + `strtok`. Original calls `set`.  |
| **`Collection::keyBy` (string)**   | ~0.4607            | ~1.2805           | **~2.8x Faster**     | Optimized path effective for simple keys.        |
| **`Collection::keyBy` (callable)** | ~0.4371            | ~0.3937           | *~1.1x Slower*       | Uses fallback. Near original performance.        |
| **`Collection::keyBy` (dot key)**  | ~2.2725            | ~2.2277           | *~1.0x (Similar)*    | Uses fallback (`data_get`). Near original performance. |

### Analysis:

*   Significant performance gains were achieved for `Arr::last`, `Arr::undot`, and `Collection::keyBy` (when used with simple string keys, a very common case).
*   A minor gain was observed for individual `Arr::set` calls.
*   The final optimized `Collection::keyBy` uses a pre-computed getter strategy. This maintains performance nearly identical to the original for callable and dot-notation keys while drastically improving the simple string key scenario.

### Safety Checks:

To verify these changes, the following tests were confirmed to pass:

*   **Arr method tests:**
    ```bash
    vendor/bin/phpunit tests/Support/SupportArrTest.php
    ```

*   **Collection::keyBy specific tests:**
    ```bash
    vendor/bin/phpunit tests/Support/SupportCollectionTest.php --filter testKeyBy
    ```


### Feedback

Any feedback is welcome if something needs improvements.