@donnysim 2025-04-22
For the `last`, why not just use `array_key_last`? All this original reverse and new pointer stuff is so unnecessary and overcomplicated.