@rodrigopedra 2025-04-22
I would consider sending only the `Arr::last()` changes, where your approach looks better than using `array_reverse()` on rather large arrays. 

I wonder how large of an array you tested with, and if the improvement would be even greater on larger arrays.

The `Arr::set()` and `Arr::undot()` are marginal gains, considering 100,000 iterations, versus the added complexity and maintenance burden.

Regarding `Collection::keyBy()`, although the performance increase seems great, it makes me wonder how common is `Collection::keyBy()` to be called, and how many times it would be called on single life-cycle to justify the trade-off of the added complexity, and almost equal logic to `data_get()` and to `EnumerateValues@valueRetriever()`, that look as duplicate at first sight.

If those small changes are so impactful, why not make them directly on `data_get()` and to `EnumerateValues@valueRetriever()`?

But that is my opinion, let's wait on the maintainers.