@cyppe 2025-04-22
Thanks for the feedback. Im going to make some changes.