@cyppe 2025-04-22
Ok, updated some code + did re-run all benchmarks with different size of the data. 

Updated feedback is welcome.